+++ CMakeLists.txt
@@ -65,7 +65,6 @@ set(XTENSOR_HEADERS
     ${XTENSOR_INCLUDE_DIR}/xtensor/xbuffer_adaptor.hpp
     ${XTENSOR_INCLUDE_DIR}/xtensor/xbuilder.hpp
     ${XTENSOR_INCLUDE_DIR}/xtensor/xcomplex.hpp
-    ${XTENSOR_INCLUDE_DIR}/xtensor/xconcepts.hpp
     ${XTENSOR_INCLUDE_DIR}/xtensor/xcontainer.hpp
     ${XTENSOR_INCLUDE_DIR}/xtensor/xcsv.hpp
     ${XTENSOR_INCLUDE_DIR}/xtensor/xdynamic_view.hpp
