+++ include/xtensor/xnorm.hpp
@@ -14,7 +14,6 @@
 #include <complex>
 #include <cstdlib>
 
-#include "xconcepts.hpp"
 #include "xmath.hpp"
 #include "xoperation.hpp"
 #include "xutils.hpp"
