+++ include/xtensor/xoptional_assembly_storage.hpp
@@ -12,7 +12,6 @@
 
 #include "xtl/xiterator_base.hpp"
 
-#include "xconcepts.hpp"
 #include "xoptional.hpp"
 #include "xsemantic.hpp"
